+++ user_mf2microformat.h
@@ -1,0 +1,19 @@
+/*
+  +----------------------------------------------------------------------+
+  | mf2: extension for PHP                                                |
+  +----------------------------------------------------------------------+
+  | Copyright © 2019 Christopher James Willcock                          |
+  +----------------------------------------------------------------------+
+  | This source file is subject to the MIT License (MIT), that is        |
+  | bundled with this package in the file LICENSE.md.                    |
+  +----------------------------------------------------------------------+
+  | Author: Christopher James Willcock <cjwillcock@ieee.org>             |
+  +----------------------------------------------------------------------+
+*/
+
+#ifndef UPHP_MF2_MICROFORMAT_H
+#define UPHP_MF2_MICROFORMAT_H
+
+PHP_METHOD( MF2Microformat, __construct );
+
+#endif /* UPHP_MF2_MICROFORMAT_H */